+++ DXEngine/Window/VEWindow.h
@@ -18,8 +18,8 @@ namespace Venus
 
 		virtual void* Handle() const = 0;
 
-		virtual size_t Width() const = 0;
-		virtual size_t Height() const = 0;
+		virtual uint32_t Width() const = 0;
+		virtual uint32_t Height() const = 0;
 		virtual float AspectRatio() const = 0;
 
 	protected:
